requirements.txt
streamlit>=1.40.0
Pillow>=11.0.0
google-genai>=1.0.0
boto3>=1.35.0
pywebview>=5.0
